Add Dapr aggregate services integration
Introduced the `AddDaprAggregateServices` method in the `DaprServicesHelper` class to register an `IAggregateService` implementation. This method is utilized in the `HexalithWebApi` class to enhance the service collection. Additional using directives have been added to support Dapr services and improve application functionality.

Add Dapr aggregate services integration

Introduced the `AddDaprAggregateServices` method in the `DaprServicesHelper` class to register an `IAggregateService` implementation. This method is utilized in the `HexalithWebApi` class to enhance the service collection. Additional using directives have been added to support Dapr services and improve application functionality.