build(packaging): update package versions and references
update the version of `Hexalith.PolymorphicSerializations` and `Hexalith.PolymorphicSerializations.CodeGenerators` from `1.6.2` to `1.6.4`.
change the `Key` property in `EntityNotFoundException` to an auto-implemented property with a public getter.
improve XML documentation for the `innerException` parameter in `InvalidCommandAggregateIdException`.
set `PrivateAssets="true"` for `Hexalith.PolymorphicSerializations.CodeGenerators` in multiple `.csproj` files to prevent exposure to referencing projects.